moved some stuff out of drx.py into utils
